+++ src/mkvtoolnix-gui/merge/tab.h
@@ -115,6 +115,7 @@ public Q_SLOTS:
   virtual void onTrackLanguageChanged(mtx::bcp47::language_c const &newLanguage);
   virtual void onDefaultTrackFlagChanged(int newValue);
   virtual void onForcedTrackFlagChanged(int newValue);
+  virtual void onTrackEnabledFlagChanged(int newValue);
   virtual void onHearingImpairedFlagChanged(int newValue);
   virtual void onVisualImpairedFlagChanged(int newValue);
   virtual void onTextDescriptionsFlagChanged(int newValue);